requirements.txt
asgiref==3.2.10
backcall==0.2.0
branca==0.4.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
djangorestframework==3.12.1
folium==0.11.0
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
numexpr==2.7.1
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.0
prompt-toolkit==3.0.8
protobuf==3.13.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
tables==3.6.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.2.5
whitenoise==5.2.0
